Refactor wire into a standalone module.
