+++ nomos-core/chain-defs/src/wire/errors.rs
